+++ src/screens/ResetPasswordScreen/ResetPasswordScreen.tsx
@@ -1,18 +1,19 @@
 import { resetPassword } from "$apis/user.api";
+import AuthForm from "$components/common/AuthForm/AuthForm";
 import { resetPasswordSchema } from "$components/common/AuthForm/AuthForm.schema";
 import StyledButton from "$components/ui/StyledButton";
 import { USER_MESSAGES } from "$utils/constant";
-import { TeddyConfig, teddyHandsUp } from "$utils/rive";
+import { teddyHandsUp } from "$utils/rive";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useMutation } from "@tanstack/react-query";
 import { isAxiosError } from "axios";
 import React from "react";
 import { Controller, useForm } from "react-hook-form";
 import { View } from "react-native";
-import { Surface, Text, TextInput } from "react-native-paper";
+import { Text, TextInput } from "react-native-paper";
 import { SafeAreaView } from "react-native-safe-area-context";
 import Toast from "react-native-toast-message";
-import Rive, { Alignment, RiveRef } from "rive-react-native";
+import { RiveRef } from "rive-react-native";
 import { z } from "zod";
 
 export interface ResetPasswordProps {
@@ -93,31 +94,15 @@ const ResetPasswordScreen = ({
         navigation.navigate("Login");
     };
 
-    const handleLoginWithGoogle = () => {
-        console.log("Login with Google");
-    };
     return (
         <SafeAreaView className="flex-1">
             <View className="flex-1 justify-center p-4">
-                <Surface
-                    className="rounded-xl p-6 relative"
-                    style={{ height: 600 }}
-                    elevation={2}
+                <AuthForm
+                    title="Đặt lại mật khẩu"
+                    navigation={navigation}
+                    route={route}
+                    riveRef={riveRef}
                 >
-                    <Rive
-                        resourceName={TeddyConfig.resourceName}
-                        stateMachineName={TeddyConfig.stateMachineName}
-                        style={{
-                            width: 300,
-                            height: 300,
-                            position: "absolute",
-                            top: -200,
-                            left: 25,
-                        }}
-                        autoplay={false}
-                        ref={riveRef}
-                        alignment={Alignment.Center}
-                    />
                     <Text className="text-2xl font-bold mb-4 mt-20 w-full text-center">
                         Đặt lại mật khẩu
                     </Text>
@@ -195,15 +180,7 @@ const ResetPasswordScreen = ({
                             Đã có tài khoản?
                         </StyledButton>
                     </View>
-                    <StyledButton
-                        className="mt-4"
-                        mode="contained"
-                        onPress={handleLoginWithGoogle}
-                        icon="google"
-                    >
-                        Tiếp tục với Google
-                    </StyledButton>
-                </Surface>
+                </AuthForm>
             </View>
         </SafeAreaView>
     );
